Remove switching executor for onReady/onCancel/onComplete (#99)
This fixes the issue with the interleaving of onReady and onMessageAndHalfClose with the Vert.x SDK. Without this commit, both onReady and onMessageAndHalfClose will set and unset the gRPC Context. Doing so mutates the Vert.x Context, which is unique per Executor and shared among executor tasks. Because the userExecutor is Vertx.executeBlocking with ordered = false, Vert.x executes the tasks in parallel, so it is possible that the execution of onMessageAndHalfClose interleaves with the ending of onReady, which detaches the context, thus causing a syscall not found issue.

Also Update to Vert.x 4.4.4