Introduce sdk-test module to provide testing utilities (#142)
* Generate admin client
* Introduce sdk-test module to provide testing utilities.
This module provides an easy to use opinionated test toolkit to get started with testing your service with a Restate container. The container is automatically configured and deployed together with the services on the local JVM, respecting the lifecycle of JUnit 5 tests. We also provide easy to use parameter injection to send requests to services. See CounterTest for an example.

Added new allowed licenses due to the testcontainers transitive dependencies.