Extract service (discovery) protocol version from the http request headers
This commit lets the Java SDK extract the accepted ServiceDiscoveryProtocolVersions
from the accept header of the discovery request and sends a correspondingly encoded
response back if the endpoint supports the service discovery protocol version.

Likewise the SDK is now extracting the service protocol version from the content type
header when the service is invoked. If the endpoint does not support the specified
protocol version, then it will reject the request with a status code 415.